backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
google-genai==1.61.0
sqlalchemy
psycopg2-binary
supabase
